Login working list working, add not working update not working, delete not working

+++ app/src/main/java/ies/murallaromana/dam/segundo/vilaourojosealejandroproyectopmdm/activities/AddEditActivity.kt
@@ -132,7 +132,7 @@ class AddEditActivity : AppCompatActivity() {
                             if (film == null) {
                                 films.add(
                                     Film(
-                                        id.toLong(),
+                                        id,
                                         title,
                                         director,
                                         language,
@@ -147,7 +147,7 @@ class AddEditActivity : AppCompatActivity() {
                             } else {
                                 val index = films.indexOf(film)
                                 val editedFilm = Film(
-                                    id.toLong(),
+                                    id,
                                     title,
                                     director,
                                     language,
